+++ wdtk-rdf/src/test/resources/Statement.rdf
@@ -1,4 +1,4 @@
-
+<http://www.wikidata.org/Q100> <http://www.wikidata.org/prop/P227> <http://www.wikidata.org/entity/statement/Q100-id111> .
 <http://www.wikidata.org/entity/statement/Q100-id111> a <http://wikiba.se/ontology#Statement> ;
 	<http://www.wikidata.org/prop/statement/P227> "TestString" ;
 	<http://www.w3.org/ns/prov#wasDerivedFrom> <http://www.wikidata.org/reference/29b7008efe33a96335e456305edfb481> ;
